Changes done to correct datatype issues in OpenAPI spec which is causing idmpotency issues

+++ plugins/modules/feoaction.py
@@ -41,7 +41,7 @@ options:
       - When C(unset), the resource will be unset on the NetScaler ADC node.
     type: str
   cachemaxage:
-    type: float
+    type: int
     description:
       - Maxage for cache extension.
   clientsidemeasurements:
